Move *App*js to the app/ directory
