+++ app/App.js
@@ -2,7 +2,7 @@
 
 import React from 'react';
 import { Provider } from 'react-redux';
-import { setupStore } from './store';
+import { setupStore } from '../store';
 import AppContainer from './AppContainer';
 
 const store = setupStore();
