+++ app/MainApp.js
@@ -6,8 +6,8 @@ import React from 'react';
 
 import { registerCoreBlocks } from '@gutenberg/blocks/library';
 
-import BlockManager from './block-management/block-manager';
-import type { BlockListType } from './block-management/block-manager';
+import BlockManager from '../block-management/block-manager';
+import type { BlockListType } from '../block-management/block-manager';
 
 type PropsType = BlockListType;
 type StateType = {};
